+++ loader/src/platform/ios/util.mm
@@ -484,6 +484,10 @@ cocos2d::CCRect geode::utils::getSafeAreaRect() {
     );
 }
 
+double geode::utils::getInputTimestamp() {
+    return [[NSProcessInfo processInfo] systemUptime];
+}
+
 bool cocos2d::CCImage::saveToFile(const char* pszFilePath, bool bIsToRGB) {
     uint8_t* data = m_pData;
     bool usePNG = std::string_view(pszFilePath).ends_with(".png");